refactor(admin): thin server component 패턴 적용
